requirements.txt
langchain
langchain-groq
langchain-community
pydantic
fastapi
fastapi[standard]
pexels-api-py
python-pptx
